scaleController/requirements.txt
paho-mqtt>=1.4.0
pyserial>=3.4
PyYAML>=5.1.2
